requirements.txt
python-telegram-bot==20.8
httpx==0.26.0
psycopg2-binary==2.9.9
anthropic==0.25.8
matplotlib==3.8.4
python-dotenv==1.0.1
Flask==3.0.2
